Use subheadings for individual items
I wanted to link to a specific item from a presentation (for LPC), and
this is not possible with the current format. Also, having an itemized
list creates an immense wall-of-text. Let's split the content into
individual subsections that have anchors.

I generally left the text without any changes except whitespace. In
some cases I fixed some obvious grammatical mistakes.